backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
requests==2.31.0

#For the real local model (uncomment when running on laptop)

transformers==4.37.0
torch==2.1.2
sentencepiece==0.1.99
google-generativeai==0.3.2
accelerate==0.26.0